% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimtime-internal.R
\name{sim.los}
\alias{sim.los}
\title{internal}
\usage{
sim.los(tpA, startstate0 = 1, absorbstate = 5, nstate = 5, ntrans = 9,
  n = 1000, L = 120)
}
\description{
internal
}
\details{
internal
}
\keyword{internal}
